fix: Set ECT(1) for DCTCP
Before this change, both Cubic and DCTCP were using ECT(0) to indicate ECN capability. In our internal setup, this caused both classic and L4S traffic to be classified as L4S in the AQM. Now, Cubic uses ECT(0) and DCTCP uses ECT(1), allowing the traffic to be correctly distinguished.